src/backend/requirements.txt
asgiref==3.9.1
Django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.14.1
